Each class has personal file. boolean fullFlag was deleted
