+++ Task4/src/task4/Apple.java
@@ -1,0 +1,19 @@
+package task4;
+
+public class Apple extends Fruit {
+    String[] color = {"red", "yellow", "green"};
+    int colorNumber;
+    public Apple () {
+
+    }
+    public Apple (int number) {
+        colorNumber = number;
+    }
+    @Override
+    public String toString() {
+        return "Type is Apple {" +
+                "weight=" + weight + "gm" +
+                ", color=" + color[colorNumber] +
+                '}';
+    }
+}
