+++ Task4/src/task4/GreenApplesFactory.java
@@ -1,0 +1,8 @@
+package task4;
+
+class GreenApplesFactory extends AbstractFabric {
+    @Override
+    public Fruit makeFruit() {
+        return new Apple(2);
+    }
+}
