+++ Task4/src/task4/Orange.java
@@ -1,0 +1,15 @@
+package task4;
+
+class Orange extends Fruit {
+    int peelThickness;
+    Orange (int peel) {
+        peelThickness = peel;
+    }
+    @Override
+    public String toString() {
+        return "Type is Orange {" +
+                "weight=" + weight + "gm"  +
+                ", peelThickness=" + peelThickness + "mm" +
+                '}';
+    }
+}
