+++ Task4/src/task4/RedApplesFactory.java
@@ -1,0 +1,8 @@
+package task4;
+
+class RedApplesFactory extends AbstractFabric {
+    @Override
+    public Fruit makeFruit() {
+        return new Apple(0);
+    }
+}
